Выделила функции в отдельный файл
