+++ Makefile
@@ -1,7 +1,9 @@
 TARGET = ./main.out
 HDRS_DIR = project/include
 
-SRCS = project/src/main.c
+SRCS = \
+		project/src/main.c \
+		project/src/utils.c
 
 .PHONY: all build rebuild check test memtest clean
