+++ project/include/utils.h
@@ -1,6 +1,23 @@
-#define UTILS_H
-#ifndef UTILS_H
+#define PROJECT_INCLUDE_UTILS_H
+#ifndef PROJECT_INCLUDE_UTILS_H
 
-#define filename "transaction.dat";
+#define filename "transaction.dat"
 
-#endif //UTILS_H
+struct masterRecord {
+	int Number;
+	char Name[20];
+	char Surname[20];
+	char addres[30];
+	char TelNumber[15];
+	double indebtedness;
+	double credit_limit;
+	double cash_payments;
+};
+
+typedef  struct  masterRecord Data;
+
+void masterWrite(FILE *ofPTR, Data Client);
+void transactionWrite(FILE *ofPTR, Data transfer);
+void blackRecord(FILE *ofPTR, FILE  *ofPTR_2, FILE *blackrecord, Data client_data, Data transfer);
+
+#endif // PROJECT_INCLUDE_UTILS_H
